requirements.txt
Flask>=2.0.0
matplotlib>=3.0.0
python-dotenv>=0.19.0
langchain>=0.0.50
langchain-community>=0.0.10
requests>=2.25.1
